services/fan/requirements.txt
RPi.GPIO==0.7.1
redis==5.2.1
python-dotenv==1.0.1
